fixed post mapping on enrolments

+++ src/main/java/SMS/schoolmanagementsystem/repositories/ModuleRepository.java
@@ -7,5 +7,5 @@ import org.springframework.data.jpa.repository.Query;
 
 import java.util.List;
 
-public interface ModuleRepository extends JpaRepository<Module, Long> {
+public interface ModuleRepository extends JpaRepository<Module, Integer> {
 }
